% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apparentTemp.R
\name{apparentTemp}
\alias{apparentTemp}
\title{Calculation of the apparent temperature.}
\usage{
apparentTemp(tas, hurs, wind)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}

\item{wind}{vector of wind at 10m in m/s.}
}
\value{
Apparent temperature in degC.
}
\description{
Calculation of the apparent temperature from temperature, relative humidity and wind.
}
\details{
Formula based on air temperature, relative humidity and wind only, as it is calculated in Steadman 1994, Buzan et al. 2015 GMD and references therein. There is a version including radiation (net radiation absorved per unit area of human body surface), but it is not implemented here.
}
\examples{
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
at <- apparentTemp(data_obs$tasmean, hurs=data_obs$hurs, wind= data_obs$wind)

}
\author{
A.Casanueva (22.03.2018).
}
