% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wbgt.Liljegren.R
\docType{data}
\name{"data_wbgt.Liljegren"}
\alias{"data_wbgt.Liljegren"}
\alias{data_wbgt.Liljegren}
\title{Example for Wet Bulb Globe Temperature in the sun (Liljegren et al. 2008).}
\format{
List of three elements:
\itemize{
\item data: numeric vector with the WBGT values in degC.
\item Tpwb: natural wet bulb temperature (Tnwb) in degC.
\item Tg: globe temperature in degC.
}
}
\usage{
data(data_wbgt.Liljegren)
}
\description{
Dataset to check the result of \code{wbgt.Liljegren} with the testing data \code{obs_data}
for summer 2003 and one station (Salamanca, Spain).
}
\keyword{datasets}
