% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wbt.Stull.R
\docType{data}
\name{"data_wbt.Stull"}
\alias{"data_wbt.Stull"}
\alias{data_wbt.Stull}
\title{Example for Wet Bulb Temperature (Stull 2011).}
\format{
Numeric vector with 92 values (days).
}
\usage{
data(data_wbt.Stull)
}
\description{
Dataset to check the result of \code{wbt.Stull} with the testing data \code{obs_data}
for summer 2003 and one station (Salamanca, Spain).
}
\keyword{datasets}
