% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt.Bernard.R
\name{wbgt.Bernard}
\alias{wbgt.Bernard}
\title{Calculation of wet bulb globe temperature, following Bernard's method.}
\usage{
wbgt.Bernard(tas, dewp, tolerance = 1e-04, noNAs = TRUE, swap = FALSE)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{dewp}{vector of dew point temperature in degC.}

\item{tolerance}{(optional): maximum final bracket width for the psychrometric
wet-bulb solution. Default: 1e-4.}

\item{noNAs}{logical, should NAs be introduced when dewp>tas? If TRUE specify how to deal in those cases (swap argument)}

\item{swap}{logical, should \code{tas >= dewp} be enforced by swapping? Otherwise, dewp is set to tas. This argument is needed when noNAs=T.}
}
\value{
A list of:

$data: wet bulb globe temperature in degC.

$Tpwb: phychrometric wet bulb temperature (Tpwb) in degC.
}
\description{
Calculation of wet bulb globe temperature from air temperature and dew point temperature. This corresponds to the implementation for indoors or shadow conditions.
}
\details{
Based on Lemke and Kjellstrom 2012, using the formulation from Bernard et al. 1999. The psychrometric wet-bulb temperature is solved with vectorized bisection on the physical interval from dew point to air temperature.
}
\examples{
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
wbgt.indoors <- wbgt.Bernard(tas=data_obs$tasmean, dewp=data_obs$dewp)

}
\author{
A.Casanueva, P. Noti, J. Bhend (21.02.2017).
}
