% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.Stull.R
\name{wbt.Stull}
\alias{wbt.Stull}
\title{Calculation of wet bulb temperature, following Stull's method.}
\usage{
wbt.Stull(tas, hurs)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}
}
\value{
Wet bulb temperature in degC.
}
\description{
Calculation of wet bulb temperature from temperature and relative humidity.
}
\details{
Formulation from Stull 2011, Journal of Applied Meteorology and Climatology.
}
\examples{
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
wbt <- wbt.Stull(data_obs$tasmean, hurs=data_obs$hurs)

}
\author{
A.Casanueva (15.08.2016).
}
